requirements.txt
flask
cryptography
Pillow